requirements.txt
torch==2.5.1
transformers==4.49.0
vllm==0.7.3
Pillow>=10.4.0
bitsandbytes==0.45.3
peft>=0.7.0
tqdm==4.67.1
numpy==1.26.4
python-dotenv>=1.0.0
gradio==5.20.1
datasets==3.3.2
pandas>=2.2.3
huggingface-hub==0.29.1
accelerate==1.4.0
unsloth==2025.2.15
trl==0.15.2